moved handles and their unit tests to /internal/app
