+++ cmd/shortener/main.go
@@ -1,30 +1,11 @@
 package main
 
 import (
-	"crypto/rand"
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
 	"github.com/go-chi/chi/v5"
+	"github.com/tempizhere/goshorty/internal/app"
 	"github.com/tempizhere/goshorty/internal/config"
-	"io"
 	"net/http"
-	"strings"
 )
-
-// Хранилище для пар "короткий ID — URL"
-var urlStore = make(map[string]string)
-
-// Создаём структуры для JSON
-type ShortenRequest struct {
-	URL string `json:"url"`
-}
-type ShortenResponse struct {
-	Result string `json:"result"`
-}
-type ExpandResponse struct {
-	URL string `json:"url"`
-}
 
 func main() {
 	// Получаем конфигурацию
@@ -35,150 +16,19 @@ func main() {
 
 	// Регистрируем обработчики
 	r.Post("/", func(w http.ResponseWriter, r *http.Request) {
-		handlePostURL(w, r, cfg)
+		app.HandlePostURL(w, r, cfg)
 	})
-	r.Get("/{id}", handleGetURL)
+	r.Get("/{id}", func(w http.ResponseWriter, r *http.Request) {
+		app.HandleGetURL(w, r)
+	})
 	r.Post("/api/shorten", func(w http.ResponseWriter, r *http.Request) {
-		handleJSONShorten(w, r, cfg)
+		app.HandleJSONShorten(w, r, cfg)
 	})
 	r.Get("/api/expand/{id}", func(w http.ResponseWriter, r *http.Request) {
-		handleJSONExpand(w, r, cfg)
+		app.HandleJSONExpand(w, r, cfg)
 	})
-
 	err := http.ListenAndServe(cfg.RunAddr, r)
 	if err != nil {
 		panic(err)
 	}
-}
-
-// Обработчик POST-запросов на "/"
-func handlePostURL(w http.ResponseWriter, r *http.Request, cfg *config.Config) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Method not allowed", http.StatusBadRequest)
-		return
-	}
-	if !strings.Contains(r.Header.Get("Content-Type"), "text/plain") {
-		http.Error(w, "Content-Type must be text/plain", http.StatusBadRequest)
-		return
-	}
-	body, err := io.ReadAll(r.Body)
-	if err != nil {
-		http.Error(w, "Failed to read request body", http.StatusBadRequest)
-		return
-	}
-	originalURL := string(body)
-	if originalURL == "" {
-		http.Error(w, "Empty URL", http.StatusBadRequest)
-		return
-	}
-	id, err := generateShortID()
-	if err != nil {
-		http.Error(w, "Failed to generate ID", http.StatusInternalServerError)
-		return
-	}
-	urlStore[id] = originalURL
-	shortURL := fmt.Sprintf("%s/%s", cfg.BaseURL, id)
-	w.Header().Set("Content-Type", "text/plain")
-	w.WriteHeader(http.StatusCreated)
-	fmt.Fprint(w, shortURL)
-}
-
-// Обработчик GET-запросов на "/{id}"
-func handleGetURL(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodGet {
-		http.Error(w, "Method not allowed", http.StatusBadRequest)
-		return
-	}
-	id := chi.URLParam(r, "id")
-	originalURL, exists := urlStore[id]
-	if !exists {
-		http.Error(w, "URL not found", http.StatusBadRequest)
-		return
-	}
-	w.Header().Set("Location", originalURL)
-	w.WriteHeader(http.StatusTemporaryRedirect)
-}
-
-// Генерирует короткий ID из URL
-func generateShortID() (string, error) {
-	bytes := make([]byte, 8)
-	_, err := rand.Read(bytes)
-	if err != nil {
-		return "", err
-	}
-	encoded := base64.URLEncoding.EncodeToString(bytes)
-	return encoded[:8], nil
-}
-
-// Обработчик POST-запросов на "/api/shorten"
-func handleJSONShorten(w http.ResponseWriter, r *http.Request, cfg *config.Config) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Method not allowed", http.StatusBadRequest)
-		return
-	}
-	if !strings.Contains(r.Header.Get("Content-Type"), "application/json") {
-		http.Error(w, "Content-Type must be application/json", http.StatusBadRequest)
-		return
-	}
-	var reqBody ShortenRequest
-	if err := json.NewDecoder(r.Body).Decode(&reqBody); err != nil {
-		http.Error(w, "Invalid JSON", http.StatusBadRequest)
-		return
-	}
-	if reqBody.URL == "" {
-		http.Error(w, "Empty URL", http.StatusBadRequest)
-		return
-	}
-	id, err := generateShortID()
-	if err != nil {
-		http.Error(w, "Failed to generate ID", http.StatusInternalServerError)
-		return
-	}
-	urlStore[id] = reqBody.URL
-	respBody := ShortenResponse{
-		Result: fmt.Sprintf("%s/%s", cfg.BaseURL, id),
-	}
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusCreated)
-	data, err := json.Marshal(respBody)
-	if err != nil {
-		http.Error(w, "Failed to encode JSON", http.StatusInternalServerError)
-		return
-	}
-	w.Write(data)
-}
-
-// Обработчик GET-запросов на "/api/expand/{id}"
-func handleJSONExpand(w http.ResponseWriter, r *http.Request, cfg *config.Config) {
-	if r.Method != http.MethodGet {
-		http.Error(w, "Method not allowed", http.StatusBadRequest)
-		return
-	}
-	id := chi.URLParam(r, "id")
-	originalURL, exists := urlStore[id]
-	if !exists {
-		respBody := struct {
-			Error string `json:"error"`
-		}{Error: "URL not found"}
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusBadRequest)
-		data, err := json.Marshal(respBody)
-		if err != nil {
-			http.Error(w, "Failed to encode JSON", http.StatusInternalServerError)
-			return
-		}
-		w.Write(data)
-		return
-	}
-	respBody := ExpandResponse{
-		URL: originalURL,
-	}
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	data, err := json.Marshal(respBody)
-	if err != nil {
-		http.Error(w, "Failed to encode JSON", http.StatusInternalServerError)
-		return
-	}
-	w.Write(data)
 }
